+++ @navikt/core/react/src/form/Fieldset/index.ts
@@ -1,2 +1,2 @@
+"use client";
 export { default as Fieldset, type FieldsetProps } from "./Fieldset";
-export { FieldsetContext } from "./context";
