+++ @navikt/core/react/src/form/checkbox/index.ts
@@ -1,3 +1,4 @@
+"use client";
 export { default as Checkbox } from "./Checkbox";
 export {
   default as CheckboxGroup,
