+++ @navikt/core/react/src/form/combobox/index.ts
@@ -1,2 +1,3 @@
+"use client";
 export { default as Combobox } from "./ComboboxProvider";
 export { type ComboboxProps } from "./types";
