+++ @navikt/core/react/src/form/confirmation-panel/confirmation-panel.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/react";
 import React, { useState } from "react";
 import { Link } from "../../link";
-import { ConfirmationPanel } from "../index";
+import { ConfirmationPanel } from "./index";
 
 export default {
   title: "ds-react/ConfirmationPanel",
