+++ @navikt/core/react/src/form/confirmation-panel/index.ts
@@ -1,0 +1,5 @@
+"use client";
+export {
+  default as ConfirmationPanel,
+  type ConfirmationPanelProps,
+} from "./ConfirmationPanel";
