+++ @navikt/core/react/src/form/error-summary/index.ts
@@ -1,2 +1,5 @@
-export { default as ErrorSummary } from "./ErrorSummary";
-export { type ErrorSummaryProps } from "./ErrorSummary";
+"use client";
+export {
+  default as ErrorSummary,
+  type ErrorSummaryProps,
+} from "./ErrorSummary";
