--- @navikt/core/react/src/form/index.ts
@@ -1,24 +1,0 @@
-export {
-  default as ConfirmationPanel,
-  type ConfirmationPanelProps,
-} from "./ConfirmationPanel";
-export { Fieldset, FieldsetContext, type FieldsetProps } from "./Fieldset";
-export { default as Select, type SelectProps } from "./Select";
-export { default as Switch, type SwitchProps } from "./Switch";
-export { default as TextField, type TextFieldProps } from "./TextField";
-export { default as Textarea, type TextareaProps } from "./Textarea";
-export {
-  Checkbox,
-  CheckboxGroup,
-  type CheckboxGroupProps,
-  type CheckboxProps,
-} from "./checkbox";
-export { Combobox as UNSAFE_Combobox, type ComboboxProps } from "./combobox";
-export { ErrorSummary, type ErrorSummaryProps } from "./error-summary";
-export {
-  Radio,
-  RadioGroup,
-  type RadioGroupProps,
-  type RadioProps,
-} from "./radio";
-export { Search, type SearchClearEvent, type SearchProps } from "./search";
