+++ @navikt/core/react/src/form/radio/index.ts
@@ -1,3 +1,4 @@
+"use client";
 export { default as Radio } from "./Radio";
 export { default as RadioGroup, type RadioGroupProps } from "./RadioGroup";
 export { type RadioProps } from "./types";
