+++ @navikt/core/react/src/form/search/index.ts
@@ -1,3 +1,4 @@
+"use client";
 export {
   default as Search,
   type SearchProps,
