+++ @navikt/core/react/src/form/search/search.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from "@storybook/react";
 import React, { useState } from "react";
-import { Search } from "../index";
+import Search from "./Search";
 
 export default {
   title: "ds-react/Search",
