+++ @navikt/core/react/src/form/select/index.ts
@@ -1,0 +1,2 @@
+"use client";
+export { default as Select, type SelectProps } from "./Select";
