+++ @navikt/core/react/src/form/switch/index.ts
@@ -1,0 +1,2 @@
+"use client";
+export { default as Switch, type SwitchProps } from "./Switch";
