+++ @navikt/core/react/src/form/textarea/index.ts
@@ -1,0 +1,2 @@
+"use client";
+export { default as Textarea, type TextareaProps } from "./Textarea";
