+++ @navikt/core/react/src/form/textfield/index.ts
@@ -1,0 +1,2 @@
+"use client";
+export { default as TextField, type TextFieldProps } from "./TextField";
